Modify: [bugs] size_t & int, makefile error

+++ so_long/srcs/check_map.c
@@ -6,7 +6,7 @@
 /*   By: jihykim2 <jihykim2@student.42seoul.kr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/07/12 17:01:18 by jihykim2          #+#    #+#             */
-/*   Updated: 2023/07/20 07:53:22 by jihykim2         ###   ########.fr       */
+/*   Updated: 2023/07/20 18:11:39 by jihykim2         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -39,7 +39,7 @@ static void	_check_value_from_map(t_maps *maps)
 	idx = 0;
 	while (maps->map[idx])
 	{
-		if (ft_strlen(maps->map[idx]) != maps->width)
+		if (ft_strlen(maps->map[idx]) != (size_t)(maps->width))
 			error_map(maps, "map: must be rectanglular\n");
 		_check_lines(maps, idx);
 		idx++;
